backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
docopt==0.6.2
idna==3.6
joblib==1.3.2
langdetect==1.0.9
nltk==3.8.1
pipreqs==0.4.13
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
yarg==0.1.9